change zodToOpenAPI
