+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/component/anyof/generate-anyof-code.ts
@@ -2,8 +2,7 @@ import type { Schema } from '../../../../../types/index.js'
 import type { Config } from '../../../../../config/index.js'
 import { getRefSchemaName } from '../../../../../core/schema/references/get-ref-schema-name.js'
 import { generateZodUnion } from '../../../../zod/generate-zod-union.js'
-import { generateZod } from '../../../../zod/generate-zod.js'
-import { generateZodToOpenAPI } from '../../../../zod-to-openapi/index.js'
+import { zodToOpenAPI } from '../../../../zod-to-openapi/index.js'
 
 /**
  * Generates the Zod code for an `anyOf` schema.
@@ -18,8 +17,8 @@ export function generateAnyOfCode(schema: Schema, config: Config): string {
   }
 
   const zodSchemas = schema.anyOf.map((subSchema) => {
-    subSchema.$ref ? getRefSchemaName(subSchema, config) : generateZodToOpenAPI(config, subSchema)
-    return generateZodToOpenAPI(config, subSchema)
+    subSchema.$ref ? getRefSchemaName(subSchema, config) : zodToOpenAPI(config, subSchema)
+    return zodToOpenAPI(config, subSchema)
   })
 
   return generateZodUnion(zodSchemas)
